Fix code formatting with forge fmt
- Apply forge fmt to all Solidity files
- Fix whitespace and indentation issues
- Ensure CI formatting checks pass

+++ script/TrustedContractExample.s.sol
@@ -13,16 +13,14 @@ import "../src/interfaces/ITrustedContractManager.sol";
 contract TrustedContractExample is Script {
     function run() external {
         console.log("=== Trusted Contract Management Example ===");
-        
+
         // Deploy all components
         SafeTxPoolCore txPoolCore = new SafeTxPoolCore();
         AddressBookManager addressBookManager = new AddressBookManager();
         DelegateCallManager delegateCallManager = new DelegateCallManager();
         TrustedContractManager trustedContractManager = new TrustedContractManager();
-        TransactionValidator transactionValidator = new TransactionValidator(
-            address(addressBookManager),
-            address(trustedContractManager)
-        );
+        TransactionValidator transactionValidator =
+            new TransactionValidator(address(addressBookManager), address(trustedContractManager));
 
         // Deploy registry
         SafeTxPoolRegistry pool = new SafeTxPoolRegistry(
@@ -56,49 +54,49 @@ contract TrustedContractExample is Script {
         vm.startPrank(safe);
 
         console.log("\n=== Adding Trusted Contracts ===");
-        
+
         // Add trusted contracts with names
         pool.addTrustedContract(safe, usdcToken, "USDC Token");
         console.log("Added USDC Token as trusted contract");
-        
+
         pool.addTrustedContract(safe, daiToken, "DAI Token");
         console.log("Added DAI Token as trusted contract");
-        
+
         pool.addTrustedContract(safe, uniswapRouter, "Uniswap V3 Router");
         console.log("Added Uniswap Router as trusted contract");
 
         console.log("\n=== Checking Trusted Status ===");
-        
+
         // Check if contracts are trusted
         console.log("USDC is trusted:", pool.isTrustedContract(safe, usdcToken));
         console.log("DAI is trusted:", pool.isTrustedContract(safe, daiToken));
         console.log("Uniswap Router is trusted:", pool.isTrustedContract(safe, uniswapRouter));
-        
+
         // Check a non-trusted contract
         address randomContract = address(0xDEADBEEF);
         console.log("Random contract is trusted:", pool.isTrustedContract(safe, randomContract));
 
         console.log("\n=== Getting All Trusted Contracts ===");
-        
+
         // Get all trusted contracts
         ITrustedContractManager.TrustedContractEntry[] memory trustedContracts = pool.getTrustedContracts(safe);
         console.log("Total trusted contracts:", trustedContracts.length);
-        
-        for (uint i = 0; i < trustedContracts.length; i++) {
+
+        for (uint256 i = 0; i < trustedContracts.length; i++) {
             console.log("Contract", i + 1, ":");
             console.log("  Address:", trustedContracts[i].contractAddress);
             console.log("  Name:", string(abi.encodePacked(trustedContracts[i].name)));
         }
 
         console.log("\n=== Updating Contract Name ===");
-        
+
         // Update a contract name
         pool.addTrustedContract(safe, uniswapRouter, "Uniswap V3 Router v2");
         console.log("Updated Uniswap Router name");
-        
+
         // Get updated list
         trustedContracts = pool.getTrustedContracts(safe);
-        for (uint i = 0; i < trustedContracts.length; i++) {
+        for (uint256 i = 0; i < trustedContracts.length; i++) {
             if (trustedContracts[i].contractAddress == uniswapRouter) {
                 console.log("Updated Uniswap Router name:", string(abi.encodePacked(trustedContracts[i].name)));
                 break;
@@ -106,14 +104,14 @@ contract TrustedContractExample is Script {
         }
 
         console.log("\n=== Removing Trusted Contract ===");
-        
+
         // Remove a trusted contract
         pool.removeTrustedContract(safe, daiToken);
         console.log("Removed DAI Token from trusted contracts");
-        
+
         // Check updated status
         console.log("DAI is still trusted:", pool.isTrustedContract(safe, daiToken));
-        
+
         // Get final list
         trustedContracts = pool.getTrustedContracts(safe);
         console.log("Final trusted contracts count:", trustedContracts.length);
